+++ applications/browser-extension/src/components/richTextEditor/toolbar/LinkButton/LinkButton.module.scss
